+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0-SNAPSHOT - 06/29/16
 
 ### Added
+- Delete ParseObject by objectId
 - Create ParseObject and save to parse-server
 - Initial commit